Added children to the renderer
Previously there was no notion of what made up a paragraph of text. It
was just an element like a heading. Now the text of a paragraph is still
accessible as an attribute on the element, but the paragraph also
contains children. Each child is a substring of the paragraph with a
type attribute. The text "this text has *emphasis* in it" will be
decomposed into "this text has ", "emphasis", and " in it".

To take advantage of this change, a handler for emphasis was also added.